Fixed old tests Integrated POM design pattern into framework Updated selenium dependency to match the latest Chrome version Added cucumber steps for UI tests

+++ src/main/java/com/endava/grocy/client/EntityClient.java
@@ -18,7 +18,7 @@ public class EntityClient extends BaseClient {
                         .post("/objects/{entity}");
     }
 
-    public Response deleteRequest(EntityType entityType, Integer productId) {
+    public Response deleteRequest(EntityType entityType, Long productId) {
         return getBasicRestConfig(EnvReader.getBasePath())
                 .pathParam("entity", entityType)
                 .pathParam("objectId", productId)
